==> 2.Rmd <==
---
title: '2'
output: html_document
date: "2022-10-19"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
suppressPackageStartupMessages({
  library(SingleCellExperiment)
  library(scater)
  library(Matrix)
  library(ggplot2)
  library(dplyr)
  library(tidyverse)
})
```

```{r}
count = read.delim("../../Data/2/exprDatRaw.tsv", sep = "\t", row.names = 1)

metadata = read.csv("../../Data/2/Human_Sample_Info.csv")

count = count[,!metadata$Treatment %in% c("Cas9_Injected_control", "CRISPR_Oct4")]
metadata = metadata[!metadata$Treatment %in% c("Cas9_Injected_control", "CRISPR_Oct4"),]

cellNames = colnames(count)
tmp = paste0("E",metadata$EmbryoDay)
timepoint = c(rep(NA,39),tmp[40:1712])

exvivo = rep("exvivo", ncol(count))
treatment = paste0(exvivo, "_", metadata$Treatment)

meta = data.frame(
  cellname = cellNames,
  timepoint = timepoint,
  celltype = metadata$Branches,
  batch_batch = NA,
  batch_embryo = metadata$Embryo,
  batch_sex = NA,
  batch_other = metadata$Stage,
  treatment = metadata$Treatment,
  type = rep("embryo", ncol(count)),
  dataset = metadata$Dataset
)
rownames(meta) = cellNames

sce = SingleCellExperiment::SingleCellExperiment(assay = list(counts = count),
                                                 colData = meta)
sce = sce[, !is.na(sce$timepoint)]
```

# QC for cells
```{r}
totalUMI <- colSums(counts(sce))
#mitochondrial
propMito <- colSums(counts(sce)[grep("^MT", rownames(sce), value = TRUE), ])/totalUMI
propRibo <- colSums(counts(sce)[grep("^RP", rownames(sce), value = TRUE), ])/totalUMI
totalGene <- colSums(counts(sce) > 0)
propZero <- colSums(counts(sce) == 0)/nrow(sce)

sce$nUMI <- totalUMI
sce$nGene <- totalGene
sce$propMT <- colSums(counts(sce)[grep("^MT", rownames(sce), value = TRUE), ])/totalUMI
sce$propRibo <- colSums(counts(sce)[grep("^RP", rownames(sce), value = TRUE), ])/totalUMI

propZeroGene <- rowSums(counts(sce) == 0)/ncol(sce)
sce = sce[propZeroGene < 1, propZero < 1]
sce = scater::logNormCounts(sce)
sce = scater::runPCA(sce, scale = TRUE)
sce = scater::runUMAP(sce, ncomponents = 5)
plotUMAP(sce, colour_by = "timepoint")

saveRDS(sce, file = "../Data/human_embryo_2.RDS")
```

```{r}
df = cbind(colData(sce), reducedDim(sce, "PCA"))
df <- as.data.frame(df)

idx = colnames(df)[8:10]
g <- lapply(idx, function(metric) {
  df$metric <- as.numeric(df[, metric])
  p <- df %>% ggplot(aes(x = PC1, y = PC2, col = metric)) +
    geom_point() 
  return(p)
})
patchwork::wrap_plots(g, ncol = 3)
```

```{r}
ggplot(df, aes(x = nGene, y = propMito, col = timepoint)) + 
  geom_point()

ggplot(df, aes(x = nGene, fill = as.factor(timepoint), alpha = 0.9)) + 
  geom_histogram(bins= 100) 
```
